COMMERCE-812 commerce-openapi-util - Extract JsonMessageBodyGenerator
